fix(services): fix all problems

+++ src/main/java/com/school/ecommerceupch/services/UserServiceImpl.java
@@ -3,16 +3,25 @@ package com.school.ecommerceupch.services;
 import com.school.ecommerceupch.controllers.dtos.requests.CreateUserRequest;
 import com.school.ecommerceupch.controllers.dtos.requests.UpdateUserRequest;
 import com.school.ecommerceupch.controllers.dtos.responses.BaseResponse;
+import com.school.ecommerceupch.controllers.dtos.responses.UserResponse;
+import com.school.ecommerceupch.controllers.exceptions.ObjectNotFoundException;
+import com.school.ecommerceupch.controllers.exceptions.UniqueConstraintViolationException;
 import com.school.ecommerceupch.entities.Role;
 import com.school.ecommerceupch.entities.User;
 import com.school.ecommerceupch.entities.pivots.UserRole;
 import com.school.ecommerceupch.repositories.IUserRepository;
+import com.school.ecommerceupch.security.UserDetailsImpl;
 import com.school.ecommerceupch.services.interfaces.IRoleService;
 import com.school.ecommerceupch.services.interfaces.IUserRoleService;
 import com.school.ecommerceupch.services.interfaces.IUserService;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class UserServiceImpl implements IUserService {
@@ -27,41 +36,39 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public BaseResponse create(CreateUserRequest request) {
-        Role defaultRole = roleService.findOneAndEnsureExistByName("ROLE_USER");
-
-        User user = repository.save(from(request));
-
-        user = addRoleToUser(defaultRole, user);
-
-        return BaseResponse.builder()
-                .data(user)
-                .message("User created correctly")
-                .success(Boolean.TRUE)
-                .httpStatus(HttpStatus.CREATED)
-                .build();
-    }
-
-    private User addRoleToUser(Role defaultRole, User user) {
-        UserRole userRole = userRoleService.create(user.getId(), defaultRole.getId());
-
-        user.getUserRoles().add(userRole);
-
-        user = repository.save(user);
-
-        return user;
-    }
-
-    @Override
     public BaseResponse get(Long id) {
 
         User user = findOneAndEnsureExistById(id);
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+        System.out.println("holaaaa " + userDetails.getFirstName() + userDetails.getLastName());
+
         return BaseResponse.builder()
-                .data(user)
+                .data(from(user))
                 .message("User found")
                 .success(Boolean.TRUE)
                 .httpStatus(HttpStatus.OK)
+                .build();
+    }
+
+    @Override
+    public BaseResponse create(CreateUserRequest request) {
+
+        if (repository.existsByEmail(request.getEmail()))
+            throw new UniqueConstraintViolationException("Email is already in use");
+
+        User user = repository.save(from(request));
+
+        Role defaultRole = roleService.findOneAndEnsureExistByName("ROLE_USER");
+
+        user = addRoleToUser(defaultRole, user);
+
+        return BaseResponse.builder()
+                .data(from(user))
+                .message("User created correctly")
+                .success(Boolean.TRUE)
+                .httpStatus(HttpStatus.CREATED)
                 .build();
     }
 
@@ -72,7 +79,7 @@ public class UserServiceImpl implements IUserService {
         user = update(user, request);
 
         return BaseResponse.builder()
-                .data(user)
+                .data(from(user))
                 .message("User updated correctly")
                 .success(Boolean.TRUE)
                 .httpStatus(HttpStatus.OK)
@@ -81,6 +88,9 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public BaseResponse delete(Long id) {
+
+        if (!repository.existsById(id))
+            throw new ObjectNotFoundException("User not found");
 
         repository.deleteById(id);
 
@@ -94,16 +104,18 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public User findOneAndEnsureExistById(Long id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+        return repository.findById(id).orElseThrow(() -> new ObjectNotFoundException("User not found"));
     }
 
-    private User from(CreateUserRequest request) {
-        User user = new User();
-        user.setEmail(request.getEmail());
-        user.setFirstName(request.getFirstName());
-        user.setLastName(request.getLastName());
-        user.setDateOfBirth(request.getDateOfBirth());
-        user.setPassword(new BCryptPasswordEncoder().encode(request.getPassword()));
+    private User addRoleToUser(Role defaultRole, User user) {
+        UserRole userRole = userRoleService.create(user, defaultRole);
+
+        List<UserRole> userRoleList = new ArrayList<>();
+        userRoleList.add(userRole);
+
+        user.setUserRoles(userRoleList);
+
+        user = repository.save(user);
 
         return user;
     }
@@ -120,5 +132,27 @@ public class UserServiceImpl implements IUserService {
         return user;
     }
 
+    private User from(CreateUserRequest request) {
+        User user = new User();
+        user.setEmail(request.getEmail());
+        user.setFirstName(request.getFirstName());
+        user.setLastName(request.getLastName());
+        user.setDateOfBirth(request.getDateOfBirth());
+        user.setPassword(new BCryptPasswordEncoder().encode(request.getPassword()));
+
+        return user;
+    }
+
+    private UserResponse from(User user) {
+        UserResponse userResponse = new UserResponse();
+
+        userResponse.setId(user.getId());
+        userResponse.setEmail(user.getEmail());
+        userResponse.setFirstName(user.getFirstName());
+        userResponse.setLastName(user.getLastName());
+        userResponse.setDateOfBirth(user.getDateOfBirth());
+
+        return userResponse;
+    }
 
 }
